requirements.txt
RPi.GPIO
pyserial
flask
